src/main/java/edu/fiuba/algo3/modelo/Obstaculos/Pozo.java
package edu.fiuba.algo3.modelo.Obstaculos;
import edu.fiuba.algo3.modelo.Mapa.Esquina;
import edu.fiuba.algo3.modelo.Vehiculos.*;

public class Pozo implements IObstaculo {

    public int penalizar(Auto auto) {
        return 3;
    }

    public int penalizar(Moto moto) {
        return 3;
    }

    @Override
    public Esquina siguienteEsquina(Esquina origen, Esquina destino, Auto auto) {
        return destino;
    }

    @Override
    public Esquina siguienteEsquina(Esquina origen, Esquina destino, Camioneta camioneta) {
        return destino;
    }

    @Override
    public Esquina siguienteEsquina(Esquina origen, Esquina destino, Moto moto) {
        return destino;
    }

    public int penalizar(Camioneta camioneta) {
        return camioneta.atravesarObstaculo(this);

    }
}
